# docs/openapi/lincloud.yml
openapi: 3.0.1
info:
  title: LinCloud API
  version: '1.0'
  description: LinCloud API to be used by SDK and Web App
  contact:
    name: Pierrick Libert
    email: pierrick.libert@codeleap.de
  license:
    name: UNLICENSED
servers:
  - url: 'https://dev.cloud.linearity.io'
    description: Development Server
  - url: 'https://stg.cloud.linearity.io'
    description: Staging Server
  - url: 'https://cloud.linearity.io'
    description: Production Server
...
components:
  securitySchemes:
    defaultUserPool:
      type: apiKey
      name: Authorization
      in: header
      x-amazon-apigateway-authtype: cognito_user_pools
      x-amazon-apigateway-authorizer:
        type: cognito_user_pools
        providerARNs:
          - value_to_be_replaced_here
    revenueCatAuthorizer:
      type: apiKey
      name: Authorization
      in: header
      x-amazon-apigateway-authorizer:
        type: token
        identityValidationExpression : value_regex_to_be_here
        authorizerResultTtlInSeconds : 60
  parameters:
...
    DeviceIdHeader:
      in: header
      name: x-device-id
      required: true
      schema:
        type: string
        pattern: '[0-9a-zA-Z\-]+'
      description: Device ID from where the request is coming from
      example: device213ABC
...
  responses:
...
    BadRequest:
      description: An object with the details of the bad parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BodyGeneral'
          examples:
            Parameters:
              summary: Bad Request Parameters
              value:
                message: Default API Gateway message
                code: BAD_REQUEST_PARAMETERS
            Body:
              summary: Bad Request Body
              value:
                message: Default API Gateway message
                code: BAD_REQUEST_BODY
            GetUser:
              summary: GET User
              value:
                message: User is already active
                code: USER_ACTIVE
            PostHierarchy:
              summary: POST Hierarchy
              value:
                message: Parent hierarchy not found
                code: HIERARCHY_PARENT_NOT_FOUND
            PatchHierarchy:
              summary: PATCH Hierarchy
              value:
                message: Filename can not empty
                code: FILE_NAME_INVALID
            PostHierarchyRecover:
              summary: POST Hierarchy Recover
              value:
                message: The parent folder was deleted preventing the recovering to happen
                code: HIERARCHY_PARENT_NOT_FOUND
            GetPolling:
              summary: GET Polling
              value:
                message: 'Invalid lastUpdated iso UTC date, got: ${lastUpdated}'
                code: POLLING_FAILED
    Unauthorized:
      description: An object requesting a token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BodyGeneral'
          examples:
            Unauthorized:
              summary: Invalid token
              value:
                message: Default API Gateway message
                code: UNAUTHORIZED
            PostAuthLogin:
              summary: POST Authentication Login
              value:
                message: Your account is pending on deletion
                code: USER_PENDING_DELETION
            PostAuthLogout:
              summary: POST Authentication Logout
              value:
                message: Your account is pending on deletion
                code: USER_PENDING_DELETION
            DeleteUser:
              summary: DELETE User
              value:
                message: User is not active
                code: USER_PENDING_DELETION
            GetUser:
              summary: GET User
              value:
                message: User is not active
                code: USER_PENDING_DELETION
            PatchUser:
              summary: PATCH User
              value:
                message: User is not active
                code: USER_PENDING_DELETION
            GetUserHistory:
              summary: GET User History
              value:
                message: User is not active
                code: USER_PENDING_DELETION
...
    NotFound:
      description: An object with the details of the resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BodyGeneral'
          examples:
            AuthLogin:
              summary: POST Authentication Login
              value:
                message: User not found
                code: USER_NOT_FOUND
            AuthLogout:
              summary: POST Authentication Logout
              value:
                message: User not found
                code: USER_NOT_FOUND
            DeleteUser:
              summary: DELETE User
              value:
                message: User not found
                code: USER_NOT_FOUND
            GetUser:
              summary: GET User
              value:
                message: User not found
                code: USER_NOT_FOUND
            PathUser:
              summary: PATCH User
              value:
                message: User not found
                code: USER_NOT_FOUND
            RecoverUser:
              summary: POST User Recovery
              value:
                message: User not found
                code: USER_NOT_FOUND
            GetUserHistory:
              summary: GET User History
              value:
                message: User not found
                code: USER_NOT_FOUND
            PostFileDowndload:
              summary: POST File Downdload
              value:
                message: File contents not found
                code: FILE_CONTENT_NOT_FOUND
            PostFileUpload:
              summary: POST File Upload
              value:
                message: File not found
                code: FILE_NOT_FOUND
            PostFileUploadComplete:
              summary: POST File Upload Complete
              value:
                message: File not found
                code: FILE_NOT_FOUND
            PostFileUploadAbort:
              summary: POST File Upload Abort
              value:
                message: File not found
                code: FILE_NOT_FOUND
            GetFile:
              summary: GET File
              value:
                message: File not found
                code: FILE_NOT_FOUND
            PostFileConflictResolve:
              summary: POST File Conflict Resolve
              value:
                message: File not found
                code: FILE_NOT_FOUND
            DeleteFileContent:
              summary: DELETE File Content
              value:
                message: File content not found
                code: FILE_CONTENT_NOT_FOUND
            PostHierarchy:
              summary: POST Hierarchy
              value:
                message: User not found
                code: USER_NOT_FOUND
            PatchHierarchyUser:
              summary: PATCH Hierarchy User Not Found
              value:
                message: User not found
                code: USER_NOT_FOUND
            PatchHierarchyHierarchy:
              summary: PATCH Hierarchy Not Found
              value:
                message: Hierarchy not found
                code: HIERARCHY_NOT_FOUND
            GetHierarchyListOwnerUser:
              summary: GET Hierarchy List Owner User Not Found
              value:
                message: User not found
                code: USER_NOT_FOUND
            GetHierarchyListOwnerHierarchy:
              summary: GET Hierarchy List Owner Hierarchy Not Found
              value:
                message: Hierarchy not found
                code: HIERARCHY_NOT_FOUND
            DeleteHierarchy:
              summary: DELETE Hierarchy
              value:
                message: Hierarchy not found
                code: HIERARCHY_NOT_FOUND
            PostHierarchyRecoverUser:
              summary: POST Hierarchy Recover User Not Found
              value:
                message: User not found
                code: USER_NOT_FOUND
            PostHierarchyRecoverHierarchy:
              summary: POST Hierarchy Recover Hierarchy Not Found
              value:
                message: Hierarchy not found
                code: HIERARCHY_NOT_FOUND
            DeleteHierarchyPermanently:
              summary: DELETE Hierarchy Permanently
              value:
                message: Hierarchy not found
                code: HIERARCHY_NOT_FOUND
            PostHierarchyBatch:
              summary: POST Hierarchy Batch
              value:
                message: User not found
                code: USER_NOT_FOUND
            GetPlan:
              summary: GET Plan
              value:
                message: Plan not found
                code: PLAN_NOT_FOUND
            GetUserPlan:
              summary: GET User Plan
              value:
                message: User plan not found
                code: USER_PLAN_NOT_FOUND
    Duplicated:
      description: An object describing the duplicated resource in DB
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BodyGeneral'
          examples:
            PostHierarchy:
              summary: POST Hierarchy
              value:
                message: Duplicated name
                code: HIERARCHY_DUPLICATED
            PatchHierarchyName:
              summary: PATCH Hierarchy Duplicated Name
              value:
                message: Duplicated name
                code: HIERARCHY_DUPLICATED
            PatchHierarchySync:
              summary: PATCH Hierarchy Synchronization
              value:
                message: Hierarchy requires synchronization
                code: HIERARCHY_REQUIRES_SYNC
            PostHierarchyRecover:
              summary: POST Hierarchy Recover
              value:
                message: Hierarchy already exists and cannot be recovered
                code: HIERARCHY_DUPLICATED
...
  schemas:
...
    BodyGeneral:
      type: object
      properties:
        message:
          type: string
        code:
          oneOf:
            - $ref: '#/components/schemas/AWSCode'
            - $ref: '#/components/schemas/CustomCode'
      example:
        message: Default message
        code: DEFAULT_CODE
    # GENERIC
...
    AWSCode:
      type: string
      enum: [ACCESS_DENIED, API_CONFIGURATION_ERROR, AUTHORIZER_CONFIGURATION_ERROR, AUTHORIZER_FAILURE, BAD_REQUEST_PARAMETERS, BAD_REQUEST_BODY, DEFAULT_4XX, DEFAULT_5XX, EXPIRED_TOKEN, INTEGRATION_FAILURE, INTEGRATION_TIMEOUT, INVALID_API_KEY, INVALID_SIGNATURE, MISSING_AUTHENTICATION_TOKEN, QUOTA_EXCEEDED, REQUEST_TOO_LARGE, RESOURCE_NOT_FOUND, THROTTLED, UNAUTHORIZED, UNSUPPORTED_MEDIA_TYPE, WAF_FILTERED]
      example: ACCESS_DENIED
    CustomCode:
      type: string
      enum: [
        AUTH_LOGIN,
        AUTH_LOGOUT,
        FILE_UPLOAD_COMPLETED,
        FILE_UPLOAD_ABORTED,
        FILE_NOT_FOUND,
        FILE_CONTENT_NOT_FOUND,
        FILE_CONTENT_DELETED,
        FILE_NAME_INVALID,
        FILE_CONFLICT_RESOLVED,
        FILE_CONFLICT_RETRIEVED,
        FILE_SIZE_EXCEEDED,
        FILE_SIZE_INVALID,
        HIERARCHY_DUPLICATED,
        HIERARCHY_NOT_FOUND,
        HIERARCHY_PARENT_NOT_FOUND,
        HIERARCHY_CREATED,
        HIERARCHY_RENAMED,
        HIERARCHY_RETRIEVED,
        HIERARCHY_DELETED,
        HIERARCHY_RECOVERED,
        HIERARCHY_PERMANENTLY_DELETED,
        HIERARCHY_SHARED_LIST_RETRIEVED,
        HIERARCHY_TRASHED_LIST_RETRIEVED,
        HIERARCHY_OWNER_LIST_RETRIEVED,
        PRESET_LIST_RETRIEVED,
        USER_CREATED,
        USER_UPDATED,
        USER_DELETED,
        USER_RECOVERED,
        USER_NOT_FOUND,
        USER_ACTIVE,
        USER_PENDING_DELETION,
        USER_PLAN_NOT_FOUND,
        USER_PLAN_RETRIEVED,
        USER_PLAN_HISTORY_RETRIEVED,
      ]
      example: AUTH_LOGIN
    # REGEX
...
paths:
...
  /hierarchy:
    parameters:
      - $ref: '#/components/parameters/DeviceIdHeader'
    post:
      operationId: post-hierarchy
      summary: CREATE Hierarchy
      description: Allows the user to create hierarchy
      tags:
        - Hierarchy
      security:
        - defaultUserPool: []
      x-amazon-apigateway-integration:
        httpMethod: POST
        type: aws_proxy
        uri: LAMBDA_INVOCATION_ARN_CREATE_HIERARCHY
      x-amazon-apigateway-request-validator: all
      requestBody:
        $ref: './components/hierarchy.yml#/components/requestBodies/PostHierarchy'
      responses:
        201:
          $ref: './components/hierarchy.yml#/components/responses/PostHierarchy'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        404:
          $ref: '#/components/responses/NotFound'
        409:
          $ref: '#/components/responses/Duplicated'
    patch:
      operationId: patch-hierarchy
      summary: PATCH Hierarchy
      description: Allows the user to rename hierarchy
      tags:
        - Hierarchy
      security:
        - defaultUserPool: []
      x-amazon-apigateway-integration:
        httpMethod: PATCH
        type: aws_proxy
        uri: LAMBDA_INVOCATION_ARN_RENAME_HIERARCHY
      x-amazon-apigateway-request-validator: all
      requestBody:
        $ref: './components/hierarchy.yml#/components/requestBodies/PatchHierarchy'
      responses:
        200:
          $ref: './components/hierarchy.yml#/components/responses/PatchHierarchy'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        404:
          $ref: '#/components/responses/NotFound'
        409:
          $ref: '#/components/responses/Duplicated'
...
  /hierarchy/{hierarchyId}:
    parameters:
      - $ref: '#/components/parameters/DeviceIdHeader'
      - $ref: './components/hierarchy.yml#/components/parameters/HierarchyIDPath'
    delete:
      operationId: delete-hierarchy
      summary: DELETE Hierarchy
      description: Allows the user to delete a list of hierarchy
      tags:
        - Hierarchy
      security:
        - defaultUserPool: []
      x-amazon-apigateway-integration:
        httpMethod: DELETE
        type: aws_proxy
        uri: LAMBDA_INVOCATION_ARN_DELETE_HIERARCHY
      x-amazon-apigateway-request-validator: all
      responses:
        200:
          $ref: './components/hierarchy.yml#/components/responses/DeleteHierarchy'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        404:
          $ref: '#/components/responses/NotFound'
...